fix: Updating utility components to be packaged.
