+++ .storybook/main.js
@@ -22,12 +22,12 @@ module.exports = {
   features: { storyStoreV7: true },
   stories: [
     // Explicitly order the main documentation.
-    // '../packages/outline-templates/default/src/.storybook/stories/guides/welcome.stories.mdx',
-    // '../packages/outline-templates/default/src/.storybook/stories/guides/development/component-development/01-main.stories.mdx',
-    //'../packages/outline-templates/default/src/.storybook/**/*.stories.@(js|jsx|ts|tsx|mdx)',
+    // '../packages/outline-storybook/default/src/.storybook/stories/guides/welcome.stories.mdx',
+    // '../packages/outline-storybook/default/src/.storybook/stories/guides/development/component-development/01-main.stories.mdx',
+    //'../packages/outline-storybook/default/src/.storybook/**/*.stories.@(js|jsx|ts|tsx|mdx)',
     //'../packages/**/src/*.stories.@(js|jsx|ts|tsx|mdx)',
     // '../packages/outline-storybook/.storybook/stories/**/*.stories.@(js|jsx|ts|tsx|mdx)',
-    '../packages/outline-storybook/stories/components/*.stories.@(js|jsx|ts|tsx|mdx)',
+    '../packages/outline-storybook/stories/components/**/*.stories.@(js|jsx|ts|tsx|mdx)',
 
     // External Stories added to outlineConfig.
     // ...includedStories,
