+++ packages/outline-demonstration/index.ts
@@ -1,0 +1,3 @@
+export { OutlineSectionLink } from './src/outline-section-link/outline-section-link';
+export { OutlineSwatch } from './src/outline-swatch/outline-swatch';
+export { OutlineSwatchSet } from './src/outline-swatch-set/outline-swatch-set';
