+++ packages/outline-demonstration/src/outline-section-link/outline-section-link.css
@@ -2,7 +2,7 @@
   @apply h-full;
 }
 
-.sb-link {
+.outline-section-link {
   @apply relative block h-full rounded-2xl shadow-xl;
 }
