+++ packages/outline-demonstration/src/outline-swatch-set/outline-swatch-set.ts
@@ -3,8 +3,9 @@ import { customElement, property } from 'lit/decorators.js';
 import { OutlineElement } from '@phase2/outline-core';
 
 import componentStyles from './outline-swatch-set.css.lit';
-import tailwindThemeConfig from '../../../resolved-tailwind-config';
-import '../outline-swatch/outline-swatch';
+//@todo: Refactor the color pallette to be more dynamic and not tied to Tailwind.
+//import tailwindThemeConfig from '../../../resolved-tailwind-config';
+import '@phase2/outline-demonstrations';
 
 export const swatchSets = [
   'outline',
