+++ packages/outline-examples/index.ts
@@ -1,0 +1,3 @@
+export { ExampleSlotControllerBase } from './src/slot-controller/example-slot-controller-base/example-slot-controller-base';
+export { ExampleSlotControllerOne } from './src/slot-controller/example-slot-controller-one/example-slot-controller-one';
+export { ExampleSlotControllerTwo } from './src/slot-controller/example-slot-controller-two/example-slot-controller-two';
