+++ packages/outline-examples/src/slot-controller/example-slot-controller-base/example-slot-controller-base.css
