+++ packages/outline-storybook/stories/components/demonstration/outline-swatch-set.stories.mdx
