+++ packages/outline-storybook/stories/guides/welcome.stories.mdx
@@ -1,7 +1,7 @@
 import { html } from 'lit';
 import { Meta, Canvas, Story, Preview } from '@storybook/addon-docs';
 
-import '../../../components/utility/sb-section-link/sb-section-link';
+import '@phase2/outline-section-link';
 
 <!-- Importing the needed components for default stories -->
 import { OutlineContainer } from '@phase2/outline-container';
@@ -39,15 +39,15 @@ import { OutlineIcon } from '@phase2/outline-icon';
 
 <outline-grid y-padding full-bleed gap-size="small">
   <outline-column col-span-default="12" col-span-md="4">
-    <sb-section-link>
+    <outline-section-link>
       <span slot="badge" class="tag">Design Tokens</span>
       <a slot="link" class="phase2-blue" href="?path=/docs/design-tokens-colors-solid--page">
         Colors
       </a>
-    </sb-section-link>
+    </outline-section-link>
   </outline-column>
   <outline-column col-span-default="12" col-span-md="4">
-    <sb-section-link>
+    <outline-section-link>
       <span slot="badge" class="tag">Icons</span>
       <a slot="link" class="dusty-blue" href="?path=/docs/media-outline-icon--an-icon">
         <outline-icon library="lucide" name="arrow-left" size="32px" class="mr-2"></outline-icon>
@@ -56,49 +56,49 @@ import { OutlineIcon } from '@phase2/outline-icon';
         <outline-icon library="lucide" name="more-vertical" size="32px" class="mr-2"></outline-icon>
         <outline-icon library="lucide" name="arrow-right" size="32px" class="mr-2"></outline-icon>
       </a>
-    </sb-section-link>
+    </outline-section-link>
   </outline-column>
   <outline-column col-span-default="12" col-span-md="4">
-    <sb-section-link>
+    <outline-section-link>
       <span slot="badge" class="tag">Design Tokens</span>
       <a slot="link" class="phase2-blue" href="?path=/docs/design-tokens-typefaces--page">
         Typography
       </a>
-    </sb-section-link>
+    </outline-section-link>
   </outline-column>
   
   <outline-column col-span-default="12">
-    <sb-section-link>
+    <outline-section-link>
       <span slot="badge" class="tag">Documentation</span>
       <a slot="link" class="medium-gray large" href="?path=/docs/documentation-guides-tooling-commands-helpful-commands--page">
         Helpful Commands
       </a>
-    </sb-section-link>
+    </outline-section-link>
   </outline-column>
 
   <outline-column col-span-default="12" col-span-md="6" col-span-lg="6">
-    <sb-section-link>
+    <outline-section-link>
       <span slot="badge" class="tag">Documentation</span>
       <a slot="link" class="karma-coral" href="?path=/docs/documentation-guides-tooling-commands-about-outline--page">
         About Outline
       </a>
-    </sb-section-link>
+    </outline-section-link>
   </outline-column>
   <outline-column row-span="2" col-span-default="12" col-span-md="6" col-span-lg="6">
-    <sb-section-link>
+    <outline-section-link>
       <span slot="badge" class="tag">Documentation</span>
       <a slot="link" class="dark-gray large" href="?path=/docs/documentation-guides-component-development-coding-guide-standards--introduction">
         Outline Coding Guide & Standards
       </a>
-    </sb-section-link>
+    </outline-section-link>
   </outline-column>
   <outline-column col-span-default="12" col-span-md="6" col-span-lg="6">
-    <sb-section-link>
+    <outline-section-link>
       <span slot="badge" class="tag">Documentation</span>
       <a slot="link" class="karma-coral" href="?path=/docs/documentation-guides-qa-uat-using-storybook--introduction">
         Using Storybook
       </a>
-    </sb-section-link>
+    </outline-section-link>
   </outline-column>
   
 </outline-grid>
