+++ packages/outline-templates/default/src/components/README.md
@@ -1,0 +1,3 @@
+# Custom Component Directory
+
+> This directory should contain the custom Web Components for your design system.
